+++ src/f_sphere.c
@@ -20,10 +20,10 @@ static coco_problem_t *sphere_problem(const size_t number_of_variables) {
   problem->problem_name = coco_strdup("sphere function");
   /* Construct a meaningful problem id */
   problem_id_length =
-      snprintf(NULL, 0, "%s_%02i", "sphere", (int)number_of_variables);
+      (size_t)snprintf(NULL, 0, "%s_%02lu", "sphere", (long)number_of_variables);
   problem->problem_id = (char *)coco_allocate_memory(problem_id_length + 1);
-  snprintf(problem->problem_id, problem_id_length + 1, "%s_%02d", "sphere",
-           (int)number_of_variables);
+  snprintf(problem->problem_id, problem_id_length + 1, "%s_%02lu", "sphere",
+           (long)number_of_variables);
 
   problem->number_of_variables = number_of_variables;
   problem->number_of_objectives = 1;
